# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/otp/send:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send a 6-digit OTP code to an email address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpSendBody"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpSendResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/otp/verify:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify a 6-digit OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyBody"
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpVerifyResponse"
        "400":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /playback:
    get:
      operationId: getPlaybackState
      tags: [tracks]
      summary: Get the saved playback state for the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Saved playback state (null if none)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackState"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: savePlaybackState
      tags: [tracks]
      summary: Save the current playback position for the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaybackStateBody"
      responses:
        "200":
          description: Playback state saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackState"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OtpSendBody:
      type: object
      properties:
        email:
          type: string
      required:
        - email
    OtpSendResponse:
      type: object
      properties:
        sent:
          type: boolean
        devCode:
          type: string
      required:
        - sent
    OtpVerifyBody:
      type: object
      properties:
        email:
          type: string
        code:
          type: string
      required:
        - email
        - code
    OtpVerifyResponse:
      type: object
      properties:
        verified:
          type: boolean
      required:
        - verified
...
    PlaybackState:
      type: object
      properties:
        trackId:
          type: ["string", "null"]
        position:
          type: number
      required:
        - trackId
        - position
    PlaybackStateBody:
      type: object
      properties:
        trackId:
          type: string
        position:
          type: number
      required:
        - trackId
        - position
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string